requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
djangorestframework==3.14.0
exceptiongroup==1.1.3
flake8==6.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
